feat(stitch-web): implement stitch-web module and refactor catalog
