--- catalog/webApp/src/webMain/resources/sqlite-worker.js
@@ -1,157 +1,0 @@
-import sqlite3InitModule from 'https://cdn.jsdelivr.net/npm/@sqlite.org/sqlite-wasm@3.46.0/index.mjs';
-
-let sqlite3 = null;
-
-// Maps to track active database connections and prepared statements by their unique IDs.
-const databases = new Map(); // stores databaseId -> SQLiteDbObject
-const statements = new Map(); // stores statementId -> SQLiteStatementObject
-
-// Counters to generate unique IDs for new database connections and statements.
-let nextDatabaseId = 0;
-let nextStatementId = 0;
-
-function openRequest(id, requestData) {
-    try {
-        const newDatabaseId = nextDatabaseId++;
-        // Use in-memory for the demo if OPFS is not available or for simplicity
-        const newDatabase = new sqlite3.oo1.DB(requestData.fileName || ':memory:');
-        databases.set(newDatabaseId, newDatabase);
-        postMessage({'id': id, data: {'databaseId': newDatabaseId}});
-    } catch (error) {
-        postMessage({'id': id, error: error.message});
-    }
-}
-
-function prepareRequest(id, requestData) {
-    try {
-        const newStatementId = nextStatementId++;
-        const resultData = {
-            'statementId': newStatementId,
-            'parameterCount': 0,
-            'columnNames': []
-        };
-        const database = databases.get(requestData.databaseId);
-        if (!database) {
-            postMessage({'id': id, error: "Invalid database ID: " + requestData.databaseId});
-            return;
-        }
-        const statement = database.prepare(requestData.sql);
-        statements.set(newStatementId, statement);
-        resultData.parameterCount = sqlite3.capi.sqlite3_bind_parameter_count(statement);
-        for (let i = 0; i < statement.columnCount; i++) {
-            resultData.columnNames.push(sqlite3.capi.sqlite3_column_name(statement, i));
-        }
-        postMessage({'id': id, data: resultData});
-    } catch (error) {
-        postMessage({'id': id, error: error.message});
-    }
-}
-
-function stepRequest(id, requestData) {
-    const statement = statements.get(requestData.statementId);
-    if (!statement) {
-        postMessage({'id': id, error: "Invalid statement ID: " + requestData.statementId});
-        return;
-    }
-    try {
-        const resultData = {
-            'rows': [],
-            'columnTypes': []
-        };
-        statement.reset()
-        statement.clearBindings()
-        for (let i = 0; i < requestData.bindings.length; i++) {
-            statement.bind(i + 1, requestData.bindings[i]);
-        }
-        while (statement.step()) {
-            if (!resultData.columnTypes.length) {
-                for (let i = 0; i < statement.columnCount; i++) {
-                    resultData.columnTypes.push(sqlite3.capi.sqlite3_column_type(statement, i));
-                }
-            }
-            resultData.rows.push(statement.get([]));
-        }
-        postMessage({'id': id, data: resultData});
-    } catch (error) {
-        postMessage({'id': id, error: error.message});
-    }
-}
-
-function closeRequest(id, requestData) {
-    if (requestData.statementId) {
-        const statement = statements.get(requestData.statementId);
-        if (!statement) {
-            postMessage({'id': id, error: "Invalid statement ID: " + requestData.statementId});
-            return;
-        }
-        try {
-            statement.finalize();
-            statements.delete(requestData.statementId);
-        } catch (error) {
-            postMessage({'id': id, error: error.message});
-        }
-    }
-
-    if (requestData.databaseId) {
-        const database = databases.get(requestData.databaseId);
-        if (!database) {
-            postMessage({'id': id, error: "Invalid database ID: " + requestData.databaseId});
-            return;
-        }
-        try {
-            database.close();
-            databases.delete(requestData.databaseId);
-        } catch (error) {
-            postMessage({'id': id, error: error.message});
-        }
-    }
-}
-
-// A map that links command names (strings) to their respective handler functions.
-const commandMap = {
-    'open': openRequest,
-    'prepare': prepareRequest,
-    'step': stepRequest,
-    'close': closeRequest,
-};
-
-function handleMessage(e) {
-    const requestMsg = e.data;
-    if (!Object.hasOwn(requestMsg, 'data') && requestMsg.data == null) {
-        postMessage(
-            {'id': requestMsg.id, 'error': "Invalid request, missing 'data'."}
-        );
-        return;
-    }
-    if (!Object.hasOwn(requestMsg.data, 'cmd') && requestMsg.data.cmd == null) {
-        postMessage(
-            {'id': requestMsg.id, 'error': "Invalid request, missing 'cmd'."}
-        );
-        return;
-    }
-    const command = requestMsg.data.cmd;
-    const requestHandler = commandMap[command];
-    if (requestHandler) {
-        requestHandler(requestMsg.id, requestMsg.data);
-    } else {
-        postMessage(
-            {'id': requestMsg.id, 'error': "Invalid request, unknown command: '" + command + "'."}
-        );
-    }
-}
-
-const messageQueue = [];
-onmessage = (e) => {
-    if (!sqlite3) {
-        messageQueue.push(e);
-    } else {
-        handleMessage(e);
-    }
-};
-
-sqlite3InitModule().then(instance => {
-    sqlite3 = instance;
-    while (messageQueue.length > 0) {
-        handleMessage(messageQueue.shift());
-    }
-});
