+++ gradle/libs.versions.toml
@@ -34,6 +34,7 @@ dokka = "2.2.0"
 spotless = "8.6.0"
 ksp = "2.3.9"
 material-icons = "1.7.3"
+kotlinx-browser = "0.5.0"
 
 [libraries]
 jetbrains-kotlin-test = { module = "org.jetbrains.kotlin:kotlin-test", version.ref = "kotlin" }
@@ -79,6 +80,7 @@ ksp-api = { module = "com.google.devtools.ksp:symbol-processing-api", version.re
 android-gradle-plugin = { group = "com.android.tools.build", name = "gradle", version.ref = "agp" }
 kotlin-gradle-plugin = { group = "org.jetbrains.kotlin", name = "kotlin-gradle-plugin", version.ref = "kotlin" }
 ksp-gradle-plugin = { group = "com.google.devtools.ksp", name = "com.google.devtools.ksp.gradle.plugin", version.ref = "ksp" }
+kotlinx-browser = { module = "org.jetbrains.kotlinx:kotlinx-browser", version.ref = "kotlinx-browser" }
 
 [plugins]
 stitch-kmp-application = { id = "stitch.kmp.application" }
